+++ sciris/__init__.py
@@ -16,6 +16,7 @@ from .sc_nested    import *
 from .sc_math      import *
 from .sc_odict     import *
 from .sc_dataframe import *
+from .sc_profiling import *
 from .sc_parallel  import *
 from .sc_fileio    import *
 from .sc_asd       import *
